requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
scikit-learn
joblib
requests
python-dotenv

